TST/REF: collect fillna tests
